+++ R/helpers.R
@@ -1,7 +1,6 @@
 ##' stop without call
 ##' @keywords internal
 ##' @noRd
-##' @param ... 
 stop2 <- function(...) {
   stop(..., call. = FALSE)
 }
@@ -47,7 +46,7 @@ remove_unwanted_vars <- function(x,
   return(draws)
 }
 ##' require package
-##' 
+##'
 ##' @param package character specifying which package is required
 ##' @param version character specifying which version is required,
 ##'   default is NULL, implying any version is acceptable