+++ R/log_lik_draws.R
@@ -27,6 +27,9 @@ log_lik_draws <- function(x, ...) {
 ##' @export
 log_lik_draws.stanfit <- function(x, joint = FALSE,
                                   log_lik_name = "log_lik", ...) {
+
+  stan_vars <- names(x)
+  log_lik_name <- stan_vars[grepl(pattern = paste0("^", log_lik_name), stan_vars)]
   log_lik <- as.array(x, pars = log_lik_name)
 
   log_lik <- posterior::as_draws_array(log_lik)
@@ -35,7 +38,7 @@ log_lik_draws.stanfit <- function(x, joint = FALSE,
     log_lik <- rowsums_draws(log_lik)
     posterior::variables(log_lik) <- log_lik_name
   }
-  
+
   return(log_lik)
 }
 
@@ -44,13 +47,15 @@ log_lik_draws.stanfit <- function(x, joint = FALSE,
 log_lik_draws.CmdStanFit <- function(x, joint = FALSE,
                                      log_lik_name = "log_lik", ...) {
 
+  stan_vars <- x$metadata()$variables
+  log_lik_name <- stan_vars[grepl(pattern = paste0("^", log_lik_name), stan_vars)]
   log_lik <- x$draws(variables = log_lik_name)
 
   if (joint) {
     log_lik <- rowsums_draws(log_lik)
     posterior::variables(log_lik) <- log_lik_name
   }
-  
+
   return(log_lik)
 }
 
@@ -59,12 +64,12 @@ log_lik_draws.CmdStanFit <- function(x, joint = FALSE,
 log_lik_draws.draws <- function(x, joint = FALSE,
                                 log_lik_name = "log_lik", ...) {
 
-  log_lik <- posterior::subset_draws(x, variable = log_lik_name)
+  log_lik <- posterior::subset_draws(x, variable = paste0("^", log_lik_name), regex = TRUE)
 
   if (joint) {
     log_lik <- rowsums_draws(log_lik)
     posterior::variables(log_lik) <- log_lik_name
   }
-  
+
   return(log_lik)
 }