+++ R/priorsense-package.R
@@ -12,7 +12,7 @@
 #'   the prior and likelihood and is the first implementation of the
 #'   method described in Kallioinen et al. (2023).
 #'
-#' 
+#'
 #'
 #' @details The main diagnostic function provided by \pkg{priorsense}
 #'   is \code{\link{powerscale_sensitivity}}. Given a fitted model or
@@ -22,11 +22,13 @@
 #'   posterior, without needing to refit the model (using Pareto
 #'   smoothed importance sampling and importance weighted moment
 #'   matching; Vehtari et al. 2022, Paananen et al. 2021).
-#' 
+#'
 #' In addition, visual diagnostics are available by first using
 #' \code{\link{powerscale_sequence}} to create a sequence of perturbed
 #' posteriors, and then a plot function such as
 #' \code{\link{powerscale_plot_ecdf}} to visualise the change.
+#'
+#' For an introduction, see `vignette("priorsense")`.
 #'
 #' The following global options are available:
 #'    * `priorsense.plot_help_text`: If `TRUE` (the default), priorsense plots will include a title and explanatory text. If `FALSE` they will not.
@@ -60,17 +62,13 @@
 
 
 #' @seealso
-#' \code{\link{powerscale_sensitivity}}
-#' \code{\link{powerscale_sequence}}
-#' \code{\link{powerscale}}
-#' \code{\link{powerscale_plot_ecdf}}
-#' \code{\link{powerscale_plot_dens}}
-#' \code{\link{powerscale_plot_quantities}}
+#' [powerscale-overview],
+#' [powerscale-plots]
 #' @template powerscale_references
 "_PACKAGE"
 
 
 ## usethis namespace: start
-##' @importFrom lifecycle deprecated
-##' ## usethis namespace: end
+#' @importFrom lifecycle deprecated
+#' ## usethis namespace: end
 #' NULL