+++ NEWS.md
@@ -1,8 +1,18 @@
+priorsense 1.2.0.9000
+---
++ add `separator` argument for specifying separator between `log_lik_name` or
+  `log_prior_name` and tags
++ fix issues affecting prior and likelihood selection
++ fix issue with length of alpha sequence and improve relevant documentation 
++ add support for `nimble` and `jagsUI` and improved related vignettes
++ allow either `colors` or `colours` to be accepted in plot functions
+
 priorsense 1.2.0
 ---
 + Allow either `variable` or `variables` to be accepted to subset
 + Fix issue with plotting of high Pareto-k values in quantities plot
 + Use new weighted ecdf from ggplot2, which fixes issue with previous implementation
++ Fix `powerscale_gradients()` handling of custom `log_prior_name` and `log_lik_name` when starting from raw draws
 
 priorsense 1.1.1
 ---